Revert "feat(dashboard): UI/UX overhaul"
